type.tsの整理
